Move `FileCodec` helpers into `extensionsv1alpha1helper` package
