+++ pkg/apis/extensions/v1alpha1/helper/filecodec.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cloudinit
+package helper
 
 import (
 	"bytes"
